add AccountActivationMailerTrait and use in AccountActivationMailer and SendConfirmationEmailController
