+++ src/User/AccountActivationMailer.php
@@ -18,6 +18,8 @@ use Symfony\Component\Translation\TranslatorInterface;
 
 class AccountActivationMailer
 {
+    use AccountActivationMailerTrait;
+
     /**
      * @var SettingsRepositoryInterface
      */
@@ -60,42 +62,9 @@ class AccountActivationMailer
             return;
         }
 
-        $data = $this->getEmailData($user, $user->email);
+        $token = $this->generateToken($user, $user->email);
+        $data = $this->getEmailData($user, $user->email, $token);
 
-        $body = $this->translator->trans('core.email.activate_account.body', $data);
-        $subject = '['.$data['{forum}'].'] '.$this->translator->trans('core.email.activate_account.subject');
-
-        $this->queue->push(new SendRawEmailJob($user->email, $subject, $body));
-    }
-
-    /**
-     * @param User $user
-     * @param string $email
-     * @return EmailToken
-     */
-    protected function generateToken(User $user, $email)
-    {
-        $token = EmailToken::generate($email, $user->id);
-        $token->save();
-
-        return $token;
-    }
-
-    /**
-     * Get the data that should be made available to email templates.
-     *
-     * @param User $user
-     * @param string $email
-     * @return array
-     */
-    protected function getEmailData(User $user, $email)
-    {
-        $token = $this->generateToken($user, $email);
-
-        return [
-            '{username}' => $user->display_name,
-            '{url}' => $this->url->to('forum')->route('confirmEmail', ['token' => $token->token]),
-            '{forum}' => $this->settings->get('forum_title')
-        ];
+        $this->sendConfirmationEmail($user, $data);
     }
 }
